Split settings-pane.tsx into focused settings section components
Extract 5 internal sub-components from settings-pane.tsx (840 → 343 lines)
into dedicated files following the existing *-settings.tsx convention:

- instance-name-settings.tsx (103 lines)
- rewrite-localhost-pins-settings.tsx (44 lines)
- copy-mode-assist-settings.tsx (105 lines)
- worktree-location-settings.tsx (98 lines)
- appearance-settings.tsx (167 lines)

Each component was already self-contained with its own state and hooks,
just co-located in the parent file. No behavior changes.
