+++ apps/web/src/components/app/settings-pane.tsx
@@ -1,5 +1,4 @@
-import { useCallback, useEffect, useRef, useState } from "react";
-import { useMutation, useQueryClient } from "@tanstack/react-query";
+import { useEffect, useState } from "react";
 import {
   ArrowDownToLine,
   Bell,
@@ -12,31 +11,26 @@ import {
 } from "lucide-react";
 
 import { AgentTypeSettings } from "@/components/app/agent-type-settings";
+import { AppearanceSettings } from "@/components/app/appearance-settings";
+import { CopyModeAssistSettings } from "@/components/app/copy-mode-assist-settings";
 import { IdeSettings } from "@/components/app/ide-settings";
+import { InstanceNameSettings } from "@/components/app/instance-name-settings";
 import { DocsContent, DOCS_SECTION_NAV } from "@/components/app/docs-pane";
 import { NotificationSettings } from "@/components/app/notification-settings";
 import { PersonalitySettings } from "@/components/app/personality-settings";
-import { RadioIndicator } from "@/components/app/radio-indicator";
 import { ReleasesAdmin } from "@/components/app/release-admin";
+import { RewriteLocalhostPinsSettings } from "@/components/app/rewrite-localhost-pins-settings";
 import { UpdatesSection } from "@/components/app/release-manager";
 import { SecuritySettings } from "@/components/app/security-settings";
 import { ServiceStatus } from "@/components/app/service-status";
 import { type ServiceState } from "@/components/app/types";
-import { Checkbox } from "@/components/ui/checkbox";
-import { type IconColorId, ICON_COLOR_OPTIONS } from "@/hooks/use-icon-color";
-import {
-  AGENT_SETTINGS_QUERY_KEY,
-  useAgentSettings,
-} from "@/hooks/use-agent-settings";
-import { useInstanceName } from "@/hooks/use-instance-name";
+import { WorktreeLocationSettings } from "@/components/app/worktree-location-settings";
+import { type IconColorId } from "@/hooks/use-icon-color";
 import { useReleaseStream } from "@/hooks/use-release-stream";
-import { useRewriteLocalhostPins } from "@/hooks/use-rewrite-localhost-pins";
-import { DEFAULT_THEME_ID, type ThemeId, THEMES } from "@/hooks/use-theme";
-import { Input } from "@/components/ui/input";
+import { type ThemeId } from "@/hooks/use-theme";
 import { type AgentType } from "@/lib/agent-types";
 import { type IdeType } from "@/lib/ide-types";
 import { api } from "@/lib/api";
-import { extractHostname } from "@/lib/rewrite-pin-url";
 import { cn } from "@/lib/utils";
 
 type SettingsSection =
@@ -68,497 +62,6 @@ const HELP_SECTION = {
   label: "Help",
   icon: BookOpenText,
 };
-
-function InstanceNameSettings(): JSX.Element {
-  const {
-    instanceName,
-    setInstanceName,
-    isSaving,
-    saveError,
-    didSave,
-    clearSaveState,
-  } = useInstanceName();
-  const [draft, setDraft] = useState(instanceName);
-  const inputRef = useRef<HTMLInputElement>(null);
-  const savedTimerRef = useRef<ReturnType<typeof setTimeout> | null>(null);
-  const [showSaved, setShowSaved] = useState(false);
-
-  // Sync draft when the stored value loads/changes (but not while the user is editing)
-  useEffect(() => {
-    if (document.activeElement !== inputRef.current) {
-      setDraft(instanceName);
-    }
-  }, [instanceName]);
-
-  // Revert draft on save error
-  useEffect(() => {
-    if (saveError) {
-      setDraft(instanceName);
-    }
-  }, [saveError, instanceName]);
-
-  // Show brief "Saved" confirmation
-  useEffect(() => {
-    if (didSave) {
-      setShowSaved(true);
-      if (savedTimerRef.current) clearTimeout(savedTimerRef.current);
-      savedTimerRef.current = setTimeout(() => {
-        setShowSaved(false);
-        clearSaveState();
-      }, 2000);
-    }
-    return () => {
-      if (savedTimerRef.current) clearTimeout(savedTimerRef.current);
-    };
-  }, [didSave, clearSaveState]);
-
-  const save = useCallback(() => {
-    const trimmed = draft.trim();
-    if (trimmed !== instanceName) {
-      setInstanceName(trimmed);
-    }
-    setDraft(trimmed);
-  }, [draft, instanceName, setInstanceName]);
-
-  return (
-    <div>
-      <label
-        htmlFor="instance-name"
-        className="mb-1.5 block text-[10px] uppercase tracking-widest text-muted-foreground"
-      >
-        Instance name
-      </label>
-      <p className="mb-3 text-sm text-muted-foreground">
-        Give this Dispatch instance a name to distinguish it from others. Shown
-        in the sidebar and browser tab.
-      </p>
-      <div className="flex items-center gap-2">
-        <Input
-          id="instance-name"
-          ref={inputRef}
-          type="text"
-          value={draft}
-          onChange={(e) => {
-            setDraft(e.target.value);
-            if (saveError) clearSaveState();
-          }}
-          onBlur={save}
-          onKeyDown={(e) => {
-            if (e.key === "Enter") {
-              save();
-              inputRef.current?.blur();
-            }
-          }}
-          disabled={isSaving}
-          placeholder="e.g. Production, Staging, Local"
-          maxLength={100}
-          className={cn("w-full max-w-sm", saveError && "border-destructive")}
-        />
-        {showSaved && !saveError ? (
-          <span className="text-xs text-muted-foreground">Saved</span>
-        ) : null}
-      </div>
-      {saveError ? (
-        <p className="mt-1.5 text-xs text-destructive">
-          Failed to save. Please try again.
-        </p>
-      ) : null}
-    </div>
-  );
-}
-
-// AppSettings (About) has been merged into UpdatesSection in release-manager.tsx
-
-function RewriteLocalhostPinsSettings(): JSX.Element {
-  const { enabled, setEnabled } = useRewriteLocalhostPins();
-  const host =
-    typeof window === "undefined" ? "" : extractHostname(window.location.host);
-
-  return (
-    <div>
-      <div className="mb-1.5 text-[10px] uppercase tracking-widest text-muted-foreground">
-        Pin links
-      </div>
-      <p className="mb-3 max-w-2xl text-sm text-muted-foreground">
-        Rewrite{" "}
-        <code className="rounded bg-muted px-1 py-0.5 font-mono text-xs">
-          localhost
-        </code>{" "}
-        in pinned URLs to{" "}
-        <code className="rounded bg-muted px-1 py-0.5 font-mono text-xs">
-          {host}
-        </code>
-        . Saved on this device.
-      </p>
-      <label
-        className={cn(
-          "flex max-w-lg items-center gap-3 rounded border border-border px-3 py-2.5 transition-colors",
-          "cursor-pointer hover:bg-muted/50"
-        )}
-      >
-        <Checkbox
-          checked={enabled}
-          onCheckedChange={(v) => setEnabled(v === true)}
-          data-testid="rewrite-localhost-pins-toggle"
-        />
-        <div className="min-w-0 text-sm font-medium text-foreground">
-          Rewrite localhost in pins
-        </div>
-      </label>
-    </div>
-  );
-}
-
-type WorktreeLocation = "sibling" | "nested";
-
-function CopyModeAssistSettings(): JSX.Element {
-  const queryClient = useQueryClient();
-  const {
-    data: agentSettings,
-    error: queryError,
-    isLoading,
-  } = useAgentSettings();
-
-  const mutation = useMutation({
-    mutationFn: (nextEnabled: boolean) =>
-      api<{ copyModeAssistEnabled: boolean }>("/api/v1/agents/settings", {
-        method: "POST",
-        body: JSON.stringify({ copyModeAssistEnabled: nextEnabled }),
-      }),
-    onMutate: async (nextEnabled) => {
-      await queryClient.cancelQueries({ queryKey: AGENT_SETTINGS_QUERY_KEY });
-      const previousSettings = queryClient.getQueryData(
-        AGENT_SETTINGS_QUERY_KEY
-      );
-      queryClient.setQueryData(AGENT_SETTINGS_QUERY_KEY, (current) => ({
-        ...(typeof current === "object" && current ? current : {}),
-        copyModeAssistEnabled: nextEnabled,
-      }));
-      return { previousSettings };
-    },
-    onError: (_error, _nextEnabled, context) => {
-      if (context?.previousSettings !== undefined) {
-        queryClient.setQueryData(
-          AGENT_SETTINGS_QUERY_KEY,
-          context.previousSettings
-        );
-      }
-    },
-    onSettled: async () => {
-      await queryClient.invalidateQueries({
-        queryKey: AGENT_SETTINGS_QUERY_KEY,
-      });
-    },
-  });
-
-  const enabled = agentSettings?.copyModeAssistEnabled ?? false;
-  const saving = mutation.isPending;
-  const errorMessage =
-    mutation.error instanceof Error
-      ? mutation.error.message
-      : queryError instanceof Error
-        ? queryError.message
-        : "";
-
-  const handleChange = useCallback(
-    async (nextEnabled: boolean) => {
-      await mutation.mutateAsync(nextEnabled);
-    },
-    [mutation]
-  );
-
-  return (
-    <div>
-      <div className="mb-1.5 text-[10px] uppercase tracking-widest text-muted-foreground">
-        Tmux scrollback assist
-      </div>
-      <p className="mb-3 max-w-2xl text-sm text-muted-foreground">
-        Detect tmux copy mode, show the scrollback banner, and add a
-        return-to-live shortcut. When off, Dispatch does not change tmux mouse
-        mode or observe copy mode at all.
-      </p>
-      <label
-        className={cn(
-          "flex max-w-xl cursor-pointer items-center gap-3 rounded border border-border px-3 py-2.5 transition-colors hover:bg-muted/50",
-          (saving || isLoading) && "pointer-events-none opacity-60"
-        )}
-      >
-        <Checkbox
-          checked={enabled}
-          onCheckedChange={(value) => void handleChange(value === true)}
-          disabled={saving || isLoading}
-          data-testid="copy-mode-assist-toggle"
-        />
-        <div className="min-w-0">
-          <div className="text-sm font-medium text-foreground">
-            Enable tmux scrollback assist
-          </div>
-          <div className="text-xs text-muted-foreground">
-            Off by default. Turn on the passive copy-mode banner and
-            return-to-live action.
-          </div>
-        </div>
-      </label>
-      {errorMessage ? (
-        <p className="mt-2 text-xs text-destructive">{errorMessage}</p>
-      ) : null}
-    </div>
-  );
-}
-
-function WorktreeLocationSettings(): JSX.Element {
-  const [worktreeLocation, setWorktreeLocation] =
-    useState<WorktreeLocation>("sibling");
-  const [saving, setSaving] = useState(false);
-
-  useEffect(() => {
-    let cancelled = false;
-    void api<{ worktreeLocation: WorktreeLocation }>("/api/v1/agents/settings")
-      .then((data) => {
-        if (!cancelled) setWorktreeLocation(data.worktreeLocation);
-      })
-      .catch(() => {});
-    return () => {
-      cancelled = true;
-    };
-  }, []);
-
-  const handleChange = useCallback(async (value: WorktreeLocation) => {
-    setWorktreeLocation(value);
-    setSaving(true);
-    try {
-      await api<{ worktreeLocation: WorktreeLocation }>(
-        "/api/v1/agents/settings",
-        {
-          method: "POST",
-          body: JSON.stringify({ worktreeLocation: value }),
-        }
-      );
-    } catch {
-      // revert on error
-    } finally {
-      setSaving(false);
-    }
-  }, []);
-
-  const options: Array<{
-    value: WorktreeLocation;
-    label: string;
-    description: string;
-  }> = [
-    {
-      value: "sibling",
-      label: "Sibling directories",
-      description:
-        "Worktrees are created next to the repo (e.g. ../repo-branch-name)",
-    },
-    {
-      value: "nested",
-      label: "Inside .dispatch/worktrees",
-      description:
-        "Worktrees are created inside the repo in .dispatch/worktrees/",
-    },
-  ];
-
-  return (
-    <div>
-      <div className="mb-1.5 text-[10px] uppercase tracking-widest text-muted-foreground">
-        Worktree location
-      </div>
-      <p className="mb-3 text-sm text-muted-foreground">
-        Choose where git worktrees are created for new agents.
-      </p>
-      <div className="grid gap-2">
-        {options.map((opt) => (
-          <button
-            key={opt.value}
-            onClick={() => void handleChange(opt.value)}
-            disabled={saving}
-            className={cn(
-              "flex items-start gap-3 rounded-md border p-3 text-left transition-colors",
-              worktreeLocation === opt.value
-                ? "border-primary bg-primary/10"
-                : "border-border hover:border-muted-foreground/30"
-            )}
-          >
-            <RadioIndicator selected={worktreeLocation === opt.value} />
-            <div className="min-w-0">
-              <div className="text-sm font-medium text-foreground">
-                {opt.label}
-              </div>
-              <div className="text-xs text-muted-foreground">
-                {opt.description}
-              </div>
-            </div>
-          </button>
-        ))}
-      </div>
-    </div>
-  );
-}
-
-function AppearanceSettings({
-  theme,
-  setTheme,
-  iconColor,
-  setIconColor,
-  isIconColorSaving,
-  iconColorError,
-  clearIconColorError,
-}: {
-  theme: ThemeId;
-  setTheme: (id: ThemeId) => void;
-  iconColor: IconColorId;
-  setIconColor: (id: IconColorId) => void;
-  isIconColorSaving: boolean;
-  iconColorError: string | null;
-  clearIconColorError: () => void;
-}): JSX.Element {
-  const [pendingColor, setPendingColor] = useState<IconColorId | null>(null);
-  const displayColor = pendingColor ?? iconColor;
-
-  // Reset optimistic state on error so the selection reverts
-  useEffect(() => {
-    if (iconColorError) {
-      setPendingColor(null);
-    }
-  }, [iconColorError]);
-
-  return (
-    <div className="flex flex-col gap-6 p-4 md:p-6">
-      <div>
-        <div className="mb-1.5 text-[10px] uppercase tracking-widest text-muted-foreground">
-          Theme
-        </div>
-        <p className="mb-3 text-sm text-muted-foreground">
-          Choose a color theme for the interface.
-        </p>
-        {(["dark", "light"] as const).map((mode) => {
-          const items = THEMES.filter((t) => t.mode === mode);
-          if (items.length === 0) return null;
-          return (
-            <div key={mode} className="mb-4 last:mb-0">
-              <div className="mb-2 text-[10px] uppercase tracking-widest text-muted-foreground/80">
-                {mode === "dark" ? "Dark" : "Light"}
-              </div>
-              <div className="grid gap-3 sm:grid-cols-2">
-                {items.map((t) => (
-                  <button
-                    key={t.id}
-                    onClick={() => setTheme(t.id)}
-                    className={cn(
-                      "flex items-start gap-3 rounded-md border p-3 text-left transition-colors",
-                      theme === t.id
-                        ? "border-primary bg-primary/10"
-                        : "border-border hover:border-muted-foreground/30"
-                    )}
-                  >
-                    <div className="mt-0.5 flex gap-1">
-                      {t.swatches.map((color, i) => (
-                        <span
-                          key={i}
-                          className="block h-4 w-4 rounded-full border border-white/10"
-                          style={{ backgroundColor: color }}
-                        />
-                      ))}
-                    </div>
-                    <div className="min-w-0 flex-1">
-                      <div className="flex items-center gap-2">
-                        <div className="text-sm font-medium text-foreground">
-                          {t.label}
-                        </div>
-                        {t.id === DEFAULT_THEME_ID && (
-                          <span className="rounded-sm border border-border bg-muted/40 px-1.5 py-0.5 text-[9px] font-medium uppercase tracking-wider text-muted-foreground">
-                            Default
-                          </span>
-                        )}
-                      </div>
-                      <div className="text-xs text-muted-foreground">
-                        {t.description}
-                      </div>
-                    </div>
-                  </button>
-                ))}
-              </div>
-            </div>
-          );
-        })}
-      </div>
-
-      <div>
-        <div className="mb-1.5 text-[10px] uppercase tracking-widest text-muted-foreground">
-          Icon Color
-        </div>
-        <p className="mb-3 text-sm text-muted-foreground">
-          Pick a color for the app icon to help distinguish multiple Dispatch
-          installations.
-        </p>
-        <div
-          className={cn(
-            "flex flex-wrap gap-2",
-            isIconColorSaving && "pointer-events-none opacity-60"
-          )}
-          role="radiogroup"
-          aria-label="Icon color"
-        >
-          {ICON_COLOR_OPTIONS.map((c) => (
-            <button
-              key={c.id}
-              role="radio"
-              aria-checked={displayColor === c.id}
-              aria-label={c.label}
-              disabled={isIconColorSaving}
-              onClick={() => {
-                if (c.id !== iconColor) {
-                  setPendingColor(c.id);
-                  setIconColor(c.id);
-                }
-              }}
-              className={cn(
-                "flex w-14 flex-col items-center gap-1 rounded-lg border-2 px-1 py-1.5 transition-all",
-                displayColor === c.id
-                  ? "border-foreground bg-foreground/10"
-                  : "border-transparent hover:border-muted-foreground/40 hover:bg-muted/30"
-              )}
-            >
-              <img
-                src={`/icons/${c.id}/brand-icon.svg`}
-                alt=""
-                className="h-7 w-7 object-contain"
-              />
-              <span
-                className={cn(
-                  "text-[10px] leading-none",
-                  displayColor === c.id
-                    ? "text-foreground"
-                    : "text-muted-foreground"
-                )}
-              >
-                {c.label}
-              </span>
-            </button>
-          ))}
-        </div>
-        {iconColorError ? (
-          <p className="mt-2 text-xs text-destructive">
-            {iconColorError}{" "}
-            <button
-              onClick={clearIconColorError}
-              className="underline hover:no-underline"
-            >
-              Dismiss
-            </button>
-          </p>
-        ) : (
-          <p className="mt-2 text-xs text-muted-foreground/70">
-            Changing the icon color will reload the page. PWA users may need to
-            reinstall for launcher icons to update.
-          </p>
-        )}
-      </div>
-    </div>
-  );
-}
 
 const ALL_VALID_SECTIONS: SettingsSection[] = [
   "general",
